Move parts of terminalService up to common/browser
